@@ -1,24 +1,9 @@
 //
 // DO NOT EDIT.
+// swift-format-ignore-file
 //
 // Generated by the protocol buffer compiler.
-// Source: echo.proto
-//
-
-//
-// Copyright 2018, gRPC Authors All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// Source: .build/checkouts/grpc-swift/Protos/examples/echo/echo.proto
 //
 import GRPC
 import NIO
@@ -136,14 +121,10 @@ extension Echo_EchoClientProtocol {
   }
 }
 
-#if compiler(>=5.6)
-@available(*, deprecated)
 extension Echo_EchoClient: @unchecked Sendable {}
-#endif // compiler(>=5.6)
 
-@available(*, deprecated, renamed: "Echo_EchoNIOClient")
 public final class Echo_EchoClient: Echo_EchoClientProtocol {
-  private let lock = Lock()
+  private let lock = NIOLock()
   private var _defaultCallOptions: CallOptions
   private var _interceptors: Echo_EchoClientInterceptorFactoryProtocol?
   public let channel: GRPCChannel
@@ -195,7 +176,6 @@ public struct Echo_EchoNIOClient: Echo_EchoClientProtocol {
   }
 }
 
-#if compiler(>=5.6)
 @available(macOS 10.15, iOS 13, tvOS 13, watchOS 6, *)
 public protocol Echo_EchoAsyncClientProtocol: GRPCClient {
   static var serviceDescriptor: GRPCServiceDescriptor { get }
@@ -367,9 +347,7 @@ public struct Echo_EchoAsyncClient: Echo_EchoAsyncClientProtocol {
   }
 }
 
-#endif // compiler(>=5.6)
-
-public protocol Echo_EchoClientInterceptorFactoryProtocol: GRPCSendable {
+public protocol Echo_EchoClientInterceptorFactoryProtocol: Sendable {
 
   /// - Returns: Interceptors to use when invoking 'get'.
   func makeGetInterceptors() -> [ClientInterceptor<Echo_EchoRequest, Echo_EchoResponse>]
@@ -420,130 +398,6 @@ public enum Echo_EchoClientMetadata {
       path: "/echo.Echo/Update",
       type: GRPCCallType.bidirectionalStreaming
     )
-  }
-}
-
-#if compiler(>=5.6)
-@available(swift, deprecated: 5.6)
-extension Echo_EchoTestClient: @unchecked Sendable {}
-#endif // compiler(>=5.6)
-
-@available(swift, deprecated: 5.6, message: "Test clients are not Sendable but the 'GRPCClient' API requires clients to be Sendable. Using a localhost client and server is the recommended alternative.")
-public final class Echo_EchoTestClient: Echo_EchoClientProtocol {
-  private let fakeChannel: FakeChannel
-  public var defaultCallOptions: CallOptions
-  public var interceptors: Echo_EchoClientInterceptorFactoryProtocol?
-
-  public var channel: GRPCChannel {
-    return self.fakeChannel
-  }
-
-  public init(
-    fakeChannel: FakeChannel = FakeChannel(),
-    defaultCallOptions callOptions: CallOptions = CallOptions(),
-    interceptors: Echo_EchoClientInterceptorFactoryProtocol? = nil
-  ) {
-    self.fakeChannel = fakeChannel
-    self.defaultCallOptions = callOptions
-    self.interceptors = interceptors
-  }
-
-  /// Make a unary response for the Get RPC. This must be called
-  /// before calling 'get'. See also 'FakeUnaryResponse'.
-  ///
-  /// - Parameter requestHandler: a handler for request parts sent by the RPC.
-  public func makeGetResponseStream(
-    _ requestHandler: @escaping (FakeRequestPart<Echo_EchoRequest>) -> () = { _ in }
-  ) -> FakeUnaryResponse<Echo_EchoRequest, Echo_EchoResponse> {
-    return self.fakeChannel.makeFakeUnaryResponse(path: Echo_EchoClientMetadata.Methods.get.path, requestHandler: requestHandler)
-  }
-
-  public func enqueueGetResponse(
-    _ response: Echo_EchoResponse,
-    _ requestHandler: @escaping (FakeRequestPart<Echo_EchoRequest>) -> () = { _ in }
-  ) {
-    let stream = self.makeGetResponseStream(requestHandler)
-    // This is the only operation on the stream; try! is fine.
-    try! stream.sendMessage(response)
-  }
-
-  /// Returns true if there are response streams enqueued for 'Get'
-  public var hasGetResponsesRemaining: Bool {
-    return self.fakeChannel.hasFakeResponseEnqueued(forPath: Echo_EchoClientMetadata.Methods.get.path)
-  }
-
-  /// Make a streaming response for the Expand RPC. This must be called
-  /// before calling 'expand'. See also 'FakeStreamingResponse'.
-  ///
-  /// - Parameter requestHandler: a handler for request parts sent by the RPC.
-  public func makeExpandResponseStream(
-    _ requestHandler: @escaping (FakeRequestPart<Echo_EchoRequest>) -> () = { _ in }
-  ) -> FakeStreamingResponse<Echo_EchoRequest, Echo_EchoResponse> {
-    return self.fakeChannel.makeFakeStreamingResponse(path: Echo_EchoClientMetadata.Methods.expand.path, requestHandler: requestHandler)
-  }
-
-  public func enqueueExpandResponses(
-    _ responses: [Echo_EchoResponse],
-    _ requestHandler: @escaping (FakeRequestPart<Echo_EchoRequest>) -> () = { _ in }
-  ) {
-    let stream = self.makeExpandResponseStream(requestHandler)
-    // These are the only operation on the stream; try! is fine.
-    responses.forEach { try! stream.sendMessage($0) }
-    try! stream.sendEnd()
-  }
-
-  /// Returns true if there are response streams enqueued for 'Expand'
-  public var hasExpandResponsesRemaining: Bool {
-    return self.fakeChannel.hasFakeResponseEnqueued(forPath: Echo_EchoClientMetadata.Methods.expand.path)
-  }
-
-  /// Make a unary response for the Collect RPC. This must be called
-  /// before calling 'collect'. See also 'FakeUnaryResponse'.
-  ///
-  /// - Parameter requestHandler: a handler for request parts sent by the RPC.
-  public func makeCollectResponseStream(
-    _ requestHandler: @escaping (FakeRequestPart<Echo_EchoRequest>) -> () = { _ in }
-  ) -> FakeUnaryResponse<Echo_EchoRequest, Echo_EchoResponse> {
-    return self.fakeChannel.makeFakeUnaryResponse(path: Echo_EchoClientMetadata.Methods.collect.path, requestHandler: requestHandler)
-  }
-
-  public func enqueueCollectResponse(
-    _ response: Echo_EchoResponse,
-    _ requestHandler: @escaping (FakeRequestPart<Echo_EchoRequest>) -> () = { _ in }
-  ) {
-    let stream = self.makeCollectResponseStream(requestHandler)
-    // This is the only operation on the stream; try! is fine.
-    try! stream.sendMessage(response)
-  }
-
-  /// Returns true if there are response streams enqueued for 'Collect'
-  public var hasCollectResponsesRemaining: Bool {
-    return self.fakeChannel.hasFakeResponseEnqueued(forPath: Echo_EchoClientMetadata.Methods.collect.path)
-  }
-
-  /// Make a streaming response for the Update RPC. This must be called
-  /// before calling 'update'. See also 'FakeStreamingResponse'.
-  ///
-  /// - Parameter requestHandler: a handler for request parts sent by the RPC.
-  public func makeUpdateResponseStream(
-    _ requestHandler: @escaping (FakeRequestPart<Echo_EchoRequest>) -> () = { _ in }
-  ) -> FakeStreamingResponse<Echo_EchoRequest, Echo_EchoResponse> {
-    return self.fakeChannel.makeFakeStreamingResponse(path: Echo_EchoClientMetadata.Methods.update.path, requestHandler: requestHandler)
-  }
-
-  public func enqueueUpdateResponses(
-    _ responses: [Echo_EchoResponse],
-    _ requestHandler: @escaping (FakeRequestPart<Echo_EchoRequest>) -> () = { _ in }
-  ) {
-    let stream = self.makeUpdateResponseStream(requestHandler)
-    // These are the only operation on the stream; try! is fine.
-    responses.forEach { try! stream.sendMessage($0) }
-    try! stream.sendEnd()
-  }
-
-  /// Returns true if there are response streams enqueued for 'Update'
-  public var hasUpdateResponsesRemaining: Bool {
-    return self.fakeChannel.hasFakeResponseEnqueued(forPath: Echo_EchoClientMetadata.Methods.update.path)
   }
 }
 
@@ -618,35 +472,33 @@ extension Echo_EchoProvider {
   }
 }
 
-#if compiler(>=5.6)
-
 /// To implement a server, implement an object which conforms to this protocol.
 @available(macOS 10.15, iOS 13, tvOS 13, watchOS 6, *)
-public protocol Echo_EchoAsyncProvider: CallHandlerProvider {
+public protocol Echo_EchoAsyncProvider: CallHandlerProvider, Sendable {
   static var serviceDescriptor: GRPCServiceDescriptor { get }
   var interceptors: Echo_EchoServerInterceptorFactoryProtocol? { get }
 
   /// Immediately returns an echo of a request.
-  @Sendable func get(
+  func get(
     request: Echo_EchoRequest,
     context: GRPCAsyncServerCallContext
   ) async throws -> Echo_EchoResponse
 
   /// Splits a request into words and returns each word in a stream of messages.
-  @Sendable func expand(
+  func expand(
     request: Echo_EchoRequest,
     responseStream: GRPCAsyncResponseStreamWriter<Echo_EchoResponse>,
     context: GRPCAsyncServerCallContext
   ) async throws
 
   /// Collects a stream of messages and returns them concatenated when the caller closes.
-  @Sendable func collect(
+  func collect(
     requestStream: GRPCAsyncRequestStream<Echo_EchoRequest>,
     context: GRPCAsyncServerCallContext
   ) async throws -> Echo_EchoResponse
 
   /// Streams back messages as they are received in an input stream.
-  @Sendable func update(
+  func update(
     requestStream: GRPCAsyncRequestStream<Echo_EchoRequest>,
     responseStream: GRPCAsyncResponseStreamWriter<Echo_EchoResponse>,
     context: GRPCAsyncServerCallContext
@@ -678,7 +530,7 @@ extension Echo_EchoAsyncProvider {
         requestDeserializer: ProtobufDeserializer<Echo_EchoRequest>(),
         responseSerializer: ProtobufSerializer<Echo_EchoResponse>(),
         interceptors: self.interceptors?.makeGetInterceptors() ?? [],
-        wrapping: self.get(request:context:)
+        wrapping: { try await self.get(request: $0, context: $1) }
       )
 
     case "Expand":
@@ -687,7 +539,7 @@ extension Echo_EchoAsyncProvider {
         requestDeserializer: ProtobufDeserializer<Echo_EchoRequest>(),
         responseSerializer: ProtobufSerializer<Echo_EchoResponse>(),
         interceptors: self.interceptors?.makeExpandInterceptors() ?? [],
-        wrapping: self.expand(request:responseStream:context:)
+        wrapping: { try await self.expand(request: $0, responseStream: $1, context: $2) }
       )
 
     case "Collect":
@@ -696,7 +548,7 @@ extension Echo_EchoAsyncProvider {
         requestDeserializer: ProtobufDeserializer<Echo_EchoRequest>(),
         responseSerializer: ProtobufSerializer<Echo_EchoResponse>(),
         interceptors: self.interceptors?.makeCollectInterceptors() ?? [],
-        wrapping: self.collect(requestStream:context:)
+        wrapping: { try await self.collect(requestStream: $0, context: $1) }
       )
 
     case "Update":
@@ -705,7 +557,7 @@ extension Echo_EchoAsyncProvider {
         requestDeserializer: ProtobufDeserializer<Echo_EchoRequest>(),
         responseSerializer: ProtobufSerializer<Echo_EchoResponse>(),
         interceptors: self.interceptors?.makeUpdateInterceptors() ?? [],
-        wrapping: self.update(requestStream:responseStream:context:)
+        wrapping: { try await self.update(requestStream: $0, responseStream: $1, context: $2) }
       )
 
     default:
@@ -714,9 +566,7 @@ extension Echo_EchoAsyncProvider {
   }
 }
 
-#endif // compiler(>=5.6)
-
-public protocol Echo_EchoServerInterceptorFactoryProtocol {
+public protocol Echo_EchoServerInterceptorFactoryProtocol: Sendable {
 
   /// - Returns: Interceptors to use when handling 'get'.
   ///   Defaults to calling `self.makeInterceptors()`.
