@@ -2,7 +2,7 @@
 // swift-format-ignore-file
 //
 // Generated by the Swift generator plugin for the protocol buffer compiler.
-// Source: echo.proto
+// Source: .build/checkouts/grpc-swift/Protos/examples/echo/echo.proto
 //
 // For information on using the generated types, please see the documentation:
 //   https://github.com/apple/swift-protobuf/
@@ -34,7 +34,7 @@ fileprivate struct _GeneratedWithProtocGenSwiftVersion: SwiftProtobuf.ProtobufAP
   typealias Version = _2
 }
 
-public struct Echo_EchoRequest {
+public struct Echo_EchoRequest: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -47,7 +47,7 @@ public struct Echo_EchoRequest {
   public init() {}
 }
 
-public struct Echo_EchoResponse {
+public struct Echo_EchoResponse: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -59,11 +59,6 @@ public struct Echo_EchoResponse {
 
   public init() {}
 }
-
-#if swift(>=5.5) && canImport(_Concurrency)
-extension Echo_EchoRequest: @unchecked Sendable {}
-extension Echo_EchoResponse: @unchecked Sendable {}
-#endif  // swift(>=5.5) && canImport(_Concurrency)
 
 // MARK: - Code below here is support for the SwiftProtobuf runtime.
 
