@@ -22,23 +22,21 @@ import XCTest
 // by starting a HBApplication with gRPC support, and then using a standard grpc-swift
 // client to connect and interact with it.
 
-final class GRPCMethodAsyncTests: XCTestCase {
+final class GRPCMethodAsyncTests: ServerTestCase {
 
     func testUnaryCall() async throws {
 
         // GIVEN an HBApplication with gRPC support and a route that echos the input
-        let app = try makeEchoApplication(port: 9010)
-        app.gRPC.onUnary("echo.Echo", "Get", requestType: Echo_EchoRequest.self) { request, context async in
-            Echo_EchoResponse.with {
-                $0.text = "Swift echo get: " + request.text
+        let app = try startServer(port: 9010, serverBuilder:  {
+            $0.onUnary("echo.Echo", "Get", requestType: Echo_EchoRequest.self) { request, context async in
+                Echo_EchoResponse.with {
+                    $0.text = "Swift echo get: " + request.text
+                }
             }
-        }
-
-        try app.start()
-        defer { app.stop() }
+        })
 
         // AND GIVEN a grpc-swift client
-        let client = makeEchoClient(app: app)
+        let client = makeAsyncGRPCClient(app: app, port: 9010)
 
         // AND GIVEN a promise that the response will be received
         let group = MultiThreadedEventLoopGroup(numberOfThreads: 1)
@@ -56,18 +54,16 @@ final class GRPCMethodAsyncTests: XCTestCase {
     func testServerStreamingCall() async throws {
 
         // GIVEN an HBApplication with gRPC support and a route that expands the input
-        let app = try makeEchoApplication(port: 9011)
-        app.gRPC.onServerStream("echo.Echo", "Expand", requestType: Echo_EchoRequest.self) { request, responseStream, _ async throws in
-            for (i, part) in request.text.components(separatedBy: " ").lazy.enumerated() {
-                try await responseStream.send(Echo_EchoResponse.with { $0.text = "Swift echo expand (\(i)): \(part)" })
+        let app = try startServer(port: 9011, serverBuilder: {
+            $0.onServerStream("echo.Echo", "Expand", requestType: Echo_EchoRequest.self) { request, responseStream, _ async throws in
+                for (i, part) in request.text.components(separatedBy: " ").lazy.enumerated() {
+                    try await responseStream.send(Echo_EchoResponse.with { $0.text = "Swift echo expand (\(i)): \(part)" })
+                }
             }
-        }
-
-        try app.start()
-        defer { app.stop() }
+        })
 
         // AND GIVEN a grpc-swift client
-        let client = makeEchoClient(app: app)
+        let client = makeAsyncGRPCClient(app: app, port: 9011)
 
         // AND GIVEN a promise that the response will be received
         let group = MultiThreadedEventLoopGroup(numberOfThreads: 1)
@@ -88,20 +84,18 @@ final class GRPCMethodAsyncTests: XCTestCase {
     func testClientStreamingCall() async throws {
 
         // GIVEN an HBApplication with gRPC support and a route that collects the input
-        let app = try makeEchoApplication(port: 9012)
-        app.gRPC.onClientStream("echo.Echo", "Collect", requestType: Echo_EchoRequest.self) { requestStream, context async throws in
-            let text = try await requestStream.reduce(into: "Swift echo collect:") { result, request in
-              result += " \(request.text)"
+        let app = try startServer(port: 9012, serverBuilder: {
+            $0.onClientStream("echo.Echo", "Collect", requestType: Echo_EchoRequest.self) { requestStream, context async throws in
+                let text = try await requestStream.reduce(into: "Swift echo collect:") { result, request in
+                    result += " \(request.text)"
+                }
+                
+                return Echo_EchoResponse.with { $0.text = text }
             }
-
-            return Echo_EchoResponse.with { $0.text = text }
-        }
-
-        try app.start()
-        defer { app.stop() }
+        })
 
         // AND GIVEN a grpc-swift client
-        let client = makeEchoClient(app: app)
+        let client = makeAsyncGRPCClient(app: app, port: 9012)
 
         // AND GIVEN a promise that the response will be received
         let group = MultiThreadedEventLoopGroup(numberOfThreads: 1)
@@ -124,21 +118,19 @@ final class GRPCMethodAsyncTests: XCTestCase {
     func testBidirectionalStreamingCall() async throws {
 
         // GIVEN an HBApplication with gRPC support and a route that echos the input
-        let app = try makeEchoApplication(port: 9013)
-        app.gRPC.onBidirectionalStream("echo.Echo", "Update", requestType: Echo_EchoRequest.self) { requestStream, responseStream, _ async throws in
-            var counter = 0
-            for try await request in requestStream {
-                let text = "Swift echo update (\(counter)): \(request.text)"
-                try await responseStream.send(Echo_EchoResponse.with { $0.text = text })
-                counter += 1
+        let app = try startServer(port: 9013, serverBuilder: {
+            $0.onBidirectionalStream("echo.Echo", "Update", requestType: Echo_EchoRequest.self) { requestStream, responseStream, _ async throws in
+                var counter = 0
+                for try await request in requestStream {
+                    let text = "Swift echo update (\(counter)): \(request.text)"
+                    try await responseStream.send(Echo_EchoResponse.with { $0.text = text })
+                    counter += 1
+                }
             }
-        }
-
-        try app.start()
-        defer { app.stop() }
+        })
 
         // AND GIVEN a grpc-swift client
-        let client = makeEchoClient(app: app)
+        let client = makeAsyncGRPCClient(app: app, port: 9013)
 
         // AND GIVEN a promise that the response will be received
         let group = MultiThreadedEventLoopGroup(numberOfThreads: 1)
@@ -158,37 +150,6 @@ final class GRPCMethodAsyncTests: XCTestCase {
         XCTAssertEqual(messages.first?.text, "Swift echo update (0): First")
         XCTAssertEqual(messages.last?.text, "Swift echo update (1): Second")
 
-    }
-
-}
-
-// MARK: - Fixtures
-
-private extension GRPCMethodAsyncTests {
-
-    func makeEchoApplication(port: Int) throws -> HBApplication {
-        let app = HBApplication(configuration: .init(address: .hostname(port: port)))
-
-        try app.gRPC.addUpgrade(
-            configuration: .init(),
-            tlsConfiguration: .makeServerConfiguration(
-                certificateChain: [
-                    .certificate(SampleCertificate.server.certificate),
-                ],
-                privateKey: .privateKey(SamplePrivateKey.server)
-            )
-        )
-        return app
-    }
-
-    func makeEchoClient(app: HBApplication) -> Echo_EchoAsyncClient {
-        let channel = ClientConnection.usingPlatformAppropriateTLS(for: app.eventLoopGroup)
-            .withTLS(trustRoots: .certificates([
-                SampleCertificate.ca.certificate,
-            ]))
-            .withTLS(certificateVerification: .fullVerification)
-            .connect(host: "localhost", port: app.configuration.address.port ?? 8080)
-        return Echo_EchoAsyncClient(channel: channel)
     }
 
 }
