@@ -21,24 +21,22 @@ import XCTest
 // by starting a HBApplication with gRPC support, and then using a standard grpc-swift
 // client to connect and interact with it.
 
-final class GRPCMethodNIOTests: XCTestCase {
+final class GRPCMethodNIOTests: ServerTestCase {
 
     func testUnaryCall() throws {
 
         // GIVEN an HBApplication with gRPC support and a route that echos the input
-        let app = try makeEchoApplication(port: 9030)
-        app.gRPC.onUnary("echo.Echo", "Get", requestType: Echo_EchoRequest.self) { request, context in
-            let response = Echo_EchoResponse.with {
-                $0.text = "Swift echo get: " + request.text
+        let app = try startServer(port: 9030, serverBuilder:  {
+            $0.onUnary("echo.Echo", "Get", requestType: Echo_EchoRequest.self) { request, context in
+                let response = Echo_EchoResponse.with {
+                    $0.text = "Swift echo get: " + request.text
+                }
+                return context.eventLoop.makeSucceededFuture(response)
             }
-            return context.eventLoop.makeSucceededFuture(response)
-        }
-
-        try app.start()
-        defer { app.stop() }
+        })
 
         // AND GIVEN a grpc-swift client
-        let client = makeEchoClient(app: app)
+        let client = makeNIOGRPCClient(app: app, port: 9030)
 
         // AND GIVEN a promise that the response will be received
         let group = MultiThreadedEventLoopGroup(numberOfThreads: 1)
@@ -67,23 +65,21 @@ final class GRPCMethodNIOTests: XCTestCase {
     func testServerStreamingCall() throws {
 
         // GIVEN an HBApplication with gRPC support and a route that expands the input
-        let app = try makeEchoApplication(port: 9031)
-        app.gRPC.onServerStream("echo.Echo", "Expand", requestType: Echo_EchoRequest.self) { request, context in
-            let responses = request.text.components(separatedBy: " ").lazy.enumerated().map { i, part in
-              Echo_EchoResponse.with {
-                $0.text = "Swift echo expand (\(i)): \(part)"
-              }
+        let app = try startServer(port: 9031, serverBuilder: {
+            $0.onServerStream("echo.Echo", "Expand", requestType: Echo_EchoRequest.self) { request, context in
+                let responses = request.text.components(separatedBy: " ").lazy.enumerated().map { i, part in
+                    Echo_EchoResponse.with {
+                        $0.text = "Swift echo expand (\(i)): \(part)"
+                    }
+                }
+                
+                context.sendResponses(responses, promise: nil)
+                return context.eventLoop.makeSucceededFuture(.ok)
             }
-
-            context.sendResponses(responses, promise: nil)
-            return context.eventLoop.makeSucceededFuture(.ok)
-        }
-
-        try app.start()
-        defer { app.stop() }
+        })
 
         // AND GIVEN a grpc-swift client
-        let client = makeEchoClient(app: app)
+        let client = makeNIOGRPCClient(app: app, port: 9031)
 
         // AND GIVEN a promise that the response will be received
         let group = MultiThreadedEventLoopGroup(numberOfThreads: 1)
@@ -121,28 +117,26 @@ final class GRPCMethodNIOTests: XCTestCase {
     func testClientStreamingCall() throws {
 
         // GIVEN an HBApplication with gRPC support and a route that collects the input
-        let app = try makeEchoApplication(port: 9032)
-        app.gRPC.onClientStream("echo.Echo", "Collect", requestType: Echo_EchoRequest.self) { context in
-            var parts: [String] = []
-            return context.eventLoop.makeSucceededFuture({ event in
-                switch event {
-                case let .message(message):
-                    parts.append(message.text)
-
-                case .end:
-                    let response = Echo_EchoResponse.with {
-                        $0.text = "Swift echo collect: " + parts.joined(separator: " ")
+        let app = try startServer(port: 9032, serverBuilder: {
+            $0.onClientStream("echo.Echo", "Collect", requestType: Echo_EchoRequest.self) { context in
+                var parts: [String] = []
+                return context.eventLoop.makeSucceededFuture({ event in
+                    switch event {
+                    case let .message(message):
+                        parts.append(message.text)
+                        
+                    case .end:
+                        let response = Echo_EchoResponse.with {
+                            $0.text = "Swift echo collect: " + parts.joined(separator: " ")
+                        }
+                        context.responsePromise.succeed(response)
                     }
-                    context.responsePromise.succeed(response)
-                }
-            })
-        }
-
-        try app.start()
-        defer { app.stop() }
+                })
+            }
+        })
 
         // AND GIVEN a grpc-swift client
-        let client = makeEchoClient(app: app)
+        let client = makeNIOGRPCClient(app: app, port: 9032)
 
         // AND GIVEN a promise that the response will be received
         let group = MultiThreadedEventLoopGroup(numberOfThreads: 1)
@@ -180,29 +174,27 @@ final class GRPCMethodNIOTests: XCTestCase {
     func testBidirectionalStreamingCall() throws {
 
         // GIVEN an HBApplication with gRPC support and a route that echos the input
-        let app = try makeEchoApplication(port: 9033)
-        app.gRPC.onBidirectionalStream("echo.Echo", "Update", requestType: Echo_EchoRequest.self) { context in
-            var count = 0
-            return context.eventLoop.makeSucceededFuture({ event in
-                switch event {
-                case let .message(message):
-                    let response = Echo_EchoResponse.with {
-                        $0.text = "Swift echo update (\(count)): \(message.text)"
+        let app = try startServer(port: 9033, serverBuilder: {
+            $0.onBidirectionalStream("echo.Echo", "Update", requestType: Echo_EchoRequest.self) { context in
+                var count = 0
+                return context.eventLoop.makeSucceededFuture({ event in
+                    switch event {
+                    case let .message(message):
+                        let response = Echo_EchoResponse.with {
+                            $0.text = "Swift echo update (\(count)): \(message.text)"
+                        }
+                        count += 1
+                        context.sendResponse(response, promise: nil)
+                        
+                    case .end:
+                        context.statusPromise.succeed(.ok)
                     }
-                    count += 1
-                    context.sendResponse(response, promise: nil)
-
-                case .end:
-                    context.statusPromise.succeed(.ok)
-                }
-            })
-        }
-
-        try app.start()
-        defer { app.stop() }
+                })
+            }
+        })
 
         // AND GIVEN a grpc-swift client
-        let client = makeEchoClient(app: app)
+        let client = makeNIOGRPCClient(app: app, port: 9033)
 
         // AND GIVEN a promise that the response will be received
         let group = MultiThreadedEventLoopGroup(numberOfThreads: 1)
@@ -248,36 +240,6 @@ final class GRPCMethodNIOTests: XCTestCase {
 }
 
 // MARK: - Fixtures
-
-private extension GRPCMethodNIOTests {
-
-    func makeEchoApplication(port: Int) throws -> HBApplication {
-        let app = HBApplication(configuration: .init(address: .hostname(port: port)))
-
-        try app.gRPC.addUpgrade(
-            configuration: .init(),
-            tlsConfiguration: .makeServerConfiguration(
-                certificateChain: [
-                    .certificate(SampleCertificate.server.certificate),
-                ],
-                privateKey: .privateKey(SamplePrivateKey.server)
-            )
-        )
-
-        return app
-    }
-
-    func makeEchoClient(app: HBApplication) -> Echo_EchoNIOClient {
-        let channel = ClientConnection.usingPlatformAppropriateTLS(for: app.eventLoopGroup)
-            .withTLS(trustRoots: .certificates([
-                SampleCertificate.ca.certificate,
-            ]))
-            .withTLS(certificateVerification: .fullVerification)
-            .connect(host: "localhost", port: app.configuration.address.port ?? 8080)
-        return Echo_EchoNIOClient(channel: channel)
-    }
-
-}
 
 private extension GRPCMethodNIOTests {
 
