@@ -1,5 +1,5 @@
 /*
- * Copyright 2022, gRPC Authors All rights reserved.
+ * Copyright 2024, gRPC Authors All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -33,49 +33,49 @@ public struct SampleCertificate {
   public static let ca = SampleCertificate(
     certificate: try! NIOSSLCertificate(bytes: .init(caCert.utf8), format: .pem),
     commonName: "some-ca",
-    notAfter: Date(timeIntervalSince1970: 1699960773)
+    notAfter: Date(timeIntervalSince1970: 1752992508)
   )
 
   public static let otherCA = SampleCertificate(
     certificate: try! NIOSSLCertificate(bytes: .init(otherCACert.utf8), format: .pem),
     commonName: "some-other-ca",
-    notAfter: Date(timeIntervalSince1970: 1699960773)
+    notAfter: Date(timeIntervalSince1970: 1752992508)
   )
 
   public static let server = SampleCertificate(
     certificate: try! NIOSSLCertificate(bytes: .init(serverCert.utf8), format: .pem),
     commonName: "localhost",
-    notAfter: Date(timeIntervalSince1970: 1699960773)
+    notAfter: Date(timeIntervalSince1970: 1752992508)
   )
 
   public static let exampleServer = SampleCertificate(
     certificate: try! NIOSSLCertificate(bytes: .init(exampleServerCert.utf8), format: .pem),
     commonName: "example.com",
-    notAfter: Date(timeIntervalSince1970: 1699960773)
+    notAfter: Date(timeIntervalSince1970: 1752992508)
   )
 
   public static let serverSignedByOtherCA = SampleCertificate(
     certificate: try! NIOSSLCertificate(bytes: .init(serverSignedByOtherCACert.utf8), format: .pem),
     commonName: "localhost",
-    notAfter: Date(timeIntervalSince1970: 1699960773)
+    notAfter: Date(timeIntervalSince1970: 1752992508)
   )
 
   public static let client = SampleCertificate(
     certificate: try! NIOSSLCertificate(bytes: .init(clientCert.utf8), format: .pem),
     commonName: "localhost",
-    notAfter: Date(timeIntervalSince1970: 1699960773)
+    notAfter: Date(timeIntervalSince1970: 1752992508)
   )
 
   public static let clientSignedByOtherCA = SampleCertificate(
     certificate: try! NIOSSLCertificate(bytes: .init(clientSignedByOtherCACert.utf8), format: .pem),
     commonName: "localhost",
-    notAfter: Date(timeIntervalSince1970: 1699960773)
+    notAfter: Date(timeIntervalSince1970: 1752992508)
   )
 
   public static let exampleServerWithExplicitCurve = SampleCertificate(
     certificate: try! NIOSSLCertificate(bytes: .init(serverExplicitCurveCert.utf8), format: .pem),
     commonName: "localhost",
-    notAfter: Date(timeIntervalSince1970: 1699960773)
+    notAfter: Date(timeIntervalSince1970: 1752992508)
   )
 }
 
@@ -106,258 +106,275 @@ public struct SamplePrivateKey {
 
 private let caCert = """
 -----BEGIN CERTIFICATE-----
-MIICoDCCAYgCCQCf2sGvEIOvlDANBgkqhkiG9w0BAQsFADASMRAwDgYDVQQDDAdz
-b21lLWNhMB4XDTIyMTExNDExMTkzM1oXDTIzMTExNDExMTkzM1owEjEQMA4GA1UE
-AwwHc29tZS1jYTCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAOjCRRWS
-ezh/izFLT4g8lAjVUGbjckCnJRPfwgWDAC1adTDi4QhzONVlmYUzUkx6VPCcVVou
-vc/WM8hHC6cRDdf/ubGRZondGw6bzaO2yqc8K6BNSvqFnkuQHRpPoSc/RKHe+qTT
-glhygm3GlAUaNl0hJpXWlLqOoIb0mn8emF7afbyyWariPPQyzY2rywPLPXipitmW
-Jw7GxVC+Q2yx5GQxPvutCdtkAsrS1AsYxpvpW+kHmtj0Dj40N7yhTz1cw2QtCD2i
-CQuk9oRwtIiJi54USy/r6oq5NOlwqHyq+DGDt5XZx1RKvGJTn3ujHPEJVipoTkdX
-/K+RpqQxJNGhyO8CAwEAATANBgkqhkiG9w0BAQsFAAOCAQEAjKgbr1hwRMUwEYLe
-AAf0ODw0iRfch9QG6qYTOFjuC3KyDKdzf52OGm49Ep4gVmgMPLcA3YGtUXd8dJSf
-WYZ0pSJYeLVJiEkkV/0gwpc+vK2OxZe1XBPbW3JN+wmRBxi3MiL7bbSvDlYIj7Dv
-8c1vs8SxE4VuSkFRrcrVV0V95xs01X1/M9aa8z9Lf59ecKytaLvKysorDrCN8nC3
-zlMDehPCLH9y4UlBqp8ClUpKk/5/P8HXr40ZGq+5TFrR80YABPSVX7krRMcxIhfu
-IFIT2yhjkxMQWj8SCDUZxamBElAXY9KHSlEv3y+teRQABBVNxslHXqLKfKTF3q4S
-tUVJuA==
+MIIDBTCCAe2gAwIBAgIUE1XhwG1p6td6PtZskTnZYjQ0DkkwDQYJKoZIhvcNAQEL
+BQAwEjEQMA4GA1UEAwwHc29tZS1jYTAeFw0yNDA3MjAwNjIxNDhaFw0yNTA3MjAw
+NjIxNDhaMBIxEDAOBgNVBAMMB3NvbWUtY2EwggEiMA0GCSqGSIb3DQEBAQUAA4IB
+DwAwggEKAoIBAQCIgtl7zQcSBlvbhOY+ZRo0iMCs2Dx9H5OD6891QOgL8h2ajtsF
+c37EUbDNTqU+V1Hzt6MnIxaEy9zHAHtHCLFIRvzlFC5eeLwLAzetfRef6VX2rBfZ
+FdQe4hp/gQMbHGfxafzLtHpklxVqv9rc7YJqWk3zEMUEpv/p011oB7oKyri5zOlB
+dsXuWJFbojmLDeA8JpkR2aw5R6OtcJHMEQSJ99bQam6RYwjEPntobvgsy/UiWkHi
+BtfgNomuPoiDwwnSjFABgfQscooIAKO2elForaZyhHLN0RRo26L35nLod6bx+fgP
+wnfc/ryNIbpZT23yOpM4xgzV5ODRD0EJQqo/AgMBAAGjUzBRMB0GA1UdDgQWBBS7
+DtcKKCDw/Jy87h7R5aEnXeUMrDAfBgNVHSMEGDAWgBS7DtcKKCDw/Jy87h7R5aEn
+XeUMrDAPBgNVHRMBAf8EBTADAQH/MA0GCSqGSIb3DQEBCwUAA4IBAQASjdxCIwq9
+mTs6TlKF/qvWOoEffFHyyJXXmIS5GO6eDdWQV2L62SKZ7G+PFMEH9upRlhZZgpw0
+xCab55a/t7CEQ+zIxHBJzICnXx3CCnPWtTLN+pbduZhQbkBOdtXZyrKjcaVWRy6O
+m7719m6v5nR3m/G06VcPEoJaySu98b3zO8St1Q0QCTIKnOnfEoxkm+ADS7xhxmuA
+nRt2HqcWTH7jh8Y7PyalyalMppGAyvzOLuyAUWlcS2Uu7HTYo1NgsK60etzqP+pw
+hPvxLjkHf5snPkyzejrltjljbJDQUFENC1zcHBDVmn2QrNwVNGyzqHtnwUGZryEN
+0PzuDoAuELcN
 -----END CERTIFICATE-----
 """
 
 private let otherCACert = """
 -----BEGIN CERTIFICATE-----
-MIICrDCCAZQCCQDQxWAzi9Y9LTANBgkqhkiG9w0BAQsFADAYMRYwFAYDVQQDDA1z
-b21lLW90aGVyLWNhMB4XDTIyMTExNDExMTkzM1oXDTIzMTExNDExMTkzM1owGDEW
-MBQGA1UEAwwNc29tZS1vdGhlci1jYTCCASIwDQYJKoZIhvcNAQEBBQADggEPADCC
-AQoCggEBANOr3vaYLkfnqk0lREo0VJD/rnUGQ6BiVtKiou0uksb9gX4oHdKlnqyi
-dvFuwaJHIzjBhdWD2EqgWwuBTB3y/UybD/ZvkojnLD+QNMnbgG5aCnO03gVlVBOf
-JggEtAEM31C7Fi6X7Gr/QwRI721+kqNSB48Rj3BT93cDW73aSeL6IZ8jlvefWYR7
-1UI3bP+4WG58PSJOhUs2edaOn0G5wRZ5LyK6A77noll90cP+CVNlqLj8HRapqhf1
-XZhGwwaEYxNV1oDroxq9mcM+6E8LdWCsdE3N4Dx6pdL0lOjwvhevZ2ct/fb31NYE
-fMstojwKf9Of5/J4kZaC1mp44IwPS00CAwEAATANBgkqhkiG9w0BAQsFAAOCAQEA
-iUuX1YYdVwqamg13eji1I9/8eMP5demBnXjM7DYP3JqDhClTYNnN8aB+o1YW51ce
-3V1FtN/f3g3YMgYB4YSOb241G9uXGCz5CwcYeBCJbUT/PdNZOrTW1EzA+gAy8GxS
-yMbK+ZrXy+7mJr79sumIO2WGk//eznvgrmlKq+eZtVf/TDTYs5TdbqI4sqoN+qPQ
-WyuBXEkU2D259VvZ+GLljVr7JCysciALKDk3QAb6cfjhFh3aOqb40m5i4Jg6g2G6
-iFS1kE3KjaWhYYn66BRVOYzfT25RkFBxxJh2Pg6DQOyVUWsWJ+VrstpQlcGMElmq
-/LaIwNYfuUNcKb90L+M6vg==
+MIIDETCCAfmgAwIBAgIUHMpMfaXJAkSFODWNLAj1aYwu6QAwDQYJKoZIhvcNAQEL
+BQAwGDEWMBQGA1UEAwwNc29tZS1vdGhlci1jYTAeFw0yNDA3MjAwNjIxNDhaFw0y
+NTA3MjAwNjIxNDhaMBgxFjAUBgNVBAMMDXNvbWUtb3RoZXItY2EwggEiMA0GCSqG
+SIb3DQEBAQUAA4IBDwAwggEKAoIBAQC5JpUdD6zV340LxbVeeMve3un2lW7ydJg4
+WIYK48wjnrdSkQxP69gTnE7x3V59MyvEbjyeFnGWp3g8UDfET2JlNLMgZsSvqHcX
+XBk22RD+V+xyX2YJy1xjeGSnWe8BjfPX4WYGOQN8CzVfLpsdUO3UUjzbN3YjYsGO
+Z9MTkK1l/09+XHRSEUV9w84kpCX1+OACdd/0LXVaxF7G8ot+9i3fVIk/ftuOqD3a
+x6huwHQcio15JiHvC2V4WzU4RXNIz4wZLZCdI5Ti7lVdlO8Rv5CW4RhbUVtSFQiB
+pBHbBadKEj59nLBLIWuDn0U1JpvHP95JqUgu2p52VCwr3RO66C9PAgMBAAGjUzBR
+MB0GA1UdDgQWBBQ1BC7u2rBWDOGmhLmXxzyNWxwxqDAfBgNVHSMEGDAWgBQ1BC7u
+2rBWDOGmhLmXxzyNWxwxqDAPBgNVHRMBAf8EBTADAQH/MA0GCSqGSIb3DQEBCwUA
+A4IBAQBesCR2MNnk8zi0kSRWIbmB60+eUTnY22lzh3g5O1+NfZFp7Dhlo2jZ2Zl9
+TkltxKEL4sBZUgpkmW7yxs6lS1KT7m797+rf6XN07YJnW5NS804FUM/n5FrjxGq/
+8dZVPxjoRQ+uGUsaVbNrLmdCuKr8TYEUvuEsroGO83u/3Ub0tgxuufFxNaLID2mG
+TGjrzhkjRh9ilFT+yh5l2XeyoeYBtKifRsklZyhSjYDCWE03I5DpB/hbF5/Oy/hF
+wF5nqXGRKrzbtiCIKVSppC5+Xo1B50hKXOqNdXYnXMXtf0jSUT+Gfe9Q+7VM5fCS
+TvnoWDRZDdkQsYgRA6lPDL1T6AgH
 -----END CERTIFICATE-----
 """
 
 private let serverCert = """
 -----BEGIN CERTIFICATE-----
-MIICmjCCAYICAQEwDQYJKoZIhvcNAQELBQAwEjEQMA4GA1UEAwwHc29tZS1jYTAe
-Fw0yMjExMTQxMTE5MzNaFw0yMzExMTQxMTE5MzNaMBQxEjAQBgNVBAMMCWxvY2Fs
-aG9zdDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBANEQLRL2oOHPHN7K
-ozmP8Ks8CcndChJrunNTJAWCiMA/HFEVVfylnQpPyc/t+W2yD+d+PsJknu2hI1O5
-o53SBsEDm1taHaCJt9ur5NKpEphzOI7VuwkgcoGqmY0Hz7GmBmbG06Z6ne8EZfg5
-a/rjxhW3GyOmIT3s9xWiU3MW7VX0PDlVmkZzVYtcSp9+AXQMDpvLK48INu1mUC6u
-1nbEzj6KuFwpU5+V1cRLHer+I9HVA7qBcgsIDDEdUDG0/l0MivAyDbNHGHDZcsfj
-jwTMsGRcd+IONItHyYb72+JBEKv3/qFAe4XIeR6iJQP4OxZ4CoxeUFgkVQwqBNd+
-1JYDuvECAwEAATANBgkqhkiG9w0BAQsFAAOCAQEASoyiLe/ak0nH5Bl7RvwAsO+Y
-J2kA/oIUCpFEmsqDxK9Nt4eGIvknGzWsfTsVM4jXQo1MYCE7XrjpG9H6fubeQvxG
-b+eU3VaHztflxLouzBIM6LnzoXnt2/rjQWIMbWZri2Gwl4incvVDLOv3jm5VD1Uw
-OePLd+DvD/NzQ4nWdqCqhZAjopEPUpOT7fP8OkJVjGddvAn/0KyXkg3tutmUMB9m
-8KctofAp1fKmd056Lgj+j6DIFDxxWEiihTO1ae8FlS4X/teeGSEVGv5M4baWRrcD
-29V9XNIbMiwCNa7DJlPpxkjHdT4KifwPDHJ92RfK54SU1k0i8LD9KByuV4av9w==
+MIIC4zCCAcugAwIBAgIBATANBgkqhkiG9w0BAQsFADASMRAwDgYDVQQDDAdzb21l
+LWNhMB4XDTI0MDcyMDA2MjE0OFoXDTI1MDcyMDA2MjE0OFowFDESMBAGA1UEAwwJ
+bG9jYWxob3N0MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA0/1V9rwz
+2eRFNwdsaVZXH6Wm0Vt8lPu7bPwO96Z4LNjbcOjPGDx2dRkMUXvnYwd1B+dQ3dq+
+5tNhpMLoCjKzifc20IQLJr1ttJtPWEn5Rov+Oz2cwlz2sXR9fPMb4dV850tqzxxO
+5biUNhuEanREiyweZ3lzg0D1VMONWjKyb3hZTYBw4v6kcH7UcheaV2nYExH6tCCO
+xuGXkSSb5KxRXcMxbNWLJSlxJXjF4JFL4H7cC5eOVSRhdHyKzvhERZLlsFnjH1+C
+f6uqu8GSoiQt26qJpVR8J01Wee1YDHlfQ8P+OM2Rdzl9Bdm77d2Fdl5U35cpenio
+gHmzIYtwsegFTQIDAQABo0IwQDAdBgNVHQ4EFgQULEfZ+uKFk9G9c7+8Y0JnOTDh
+FQwwHwYDVR0jBBgwFoAUuw7XCigg8PycvO4e0eWhJ13lDKwwDQYJKoZIhvcNAQEL
+BQADggEBABfmL8OaPmYnKve73x+h3i4e4D5Oc8i9I0EQ3sAnt2+zKcOVEBHcANb5
+hjilGEqYZ7YWeI/ll6kYQUoj/ccGvqBmZRv8y/i5kyIZGRixDo4255Qg66Ubl7l5
+4hFs8dfux86sBlrWQHnR42fQ6AYax3zzWp40vfDuXiiVsN5DJ5lFHWKy4hV15JDp
+Px6jHH2JJL774MfXJhOFzfBh889g5bVlfUvs5I9kyejxDGII6bkUqyn9ivY5XODO
+It9dgfbcCl8bXBz7hCThJs+o4gJTfUBO5YI9Kq3UYonks9I2eJ0H7pYxArG0m1Wu
+4lhRYFkfhXMNnSutM8zzNSnjOAQHnwM=
 -----END CERTIFICATE-----
 """
 
 private let serverSignedByOtherCACert = """
 -----BEGIN CERTIFICATE-----
-MIICoDCCAYgCAQEwDQYJKoZIhvcNAQELBQAwGDEWMBQGA1UEAwwNc29tZS1vdGhl
-ci1jYTAeFw0yMjExMTQxMTE5MzNaFw0yMzExMTQxMTE5MzNaMBQxEjAQBgNVBAMM
-CWxvY2FsaG9zdDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBANEQLRL2
-oOHPHN7KozmP8Ks8CcndChJrunNTJAWCiMA/HFEVVfylnQpPyc/t+W2yD+d+PsJk
-nu2hI1O5o53SBsEDm1taHaCJt9ur5NKpEphzOI7VuwkgcoGqmY0Hz7GmBmbG06Z6
-ne8EZfg5a/rjxhW3GyOmIT3s9xWiU3MW7VX0PDlVmkZzVYtcSp9+AXQMDpvLK48I
-Nu1mUC6u1nbEzj6KuFwpU5+V1cRLHer+I9HVA7qBcgsIDDEdUDG0/l0MivAyDbNH
-GHDZcsfjjwTMsGRcd+IONItHyYb72+JBEKv3/qFAe4XIeR6iJQP4OxZ4CoxeUFgk
-VQwqBNd+1JYDuvECAwEAATANBgkqhkiG9w0BAQsFAAOCAQEAZN5RQsfPP09YIfYo
-UGu9m5+lpzYhE0S2+szysTg2IpWug0ZK4xhnqQYd9cGRks+U6hiLPdiyHCwOykf6
-OplIp5fMxPWZipREb9nA33Ra1G9vpB/tZxQJxDTvUeCH88SQOszdZk79+zWyVkaF
-+TCa3jDXb/vT20+wKxpPUjse5w2j0VOh21KaP82EMyOY/ZvhbMC60QyHnFDvJAEV
-sle77vbdLjYELpYUpf9N+TxFDZ2B4dY/edprLZGt3LcUUFv/WB8FxZdWcjdZML2F
-TMqicbP7H27+V1HF1rFUJWKzDNh4Wg6bY6lQNTZeHUyLwf/WlUraXTKYpqSH8FQ1
-703RGQ==
+MIIC6TCCAdGgAwIBAgIBATANBgkqhkiG9w0BAQsFADAYMRYwFAYDVQQDDA1zb21l
+LW90aGVyLWNhMB4XDTI0MDcyMDA2MjE0OFoXDTI1MDcyMDA2MjE0OFowFDESMBAG
+A1UEAwwJbG9jYWxob3N0MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA
+0/1V9rwz2eRFNwdsaVZXH6Wm0Vt8lPu7bPwO96Z4LNjbcOjPGDx2dRkMUXvnYwd1
+B+dQ3dq+5tNhpMLoCjKzifc20IQLJr1ttJtPWEn5Rov+Oz2cwlz2sXR9fPMb4dV8
+50tqzxxO5biUNhuEanREiyweZ3lzg0D1VMONWjKyb3hZTYBw4v6kcH7UcheaV2nY
+ExH6tCCOxuGXkSSb5KxRXcMxbNWLJSlxJXjF4JFL4H7cC5eOVSRhdHyKzvhERZLl
+sFnjH1+Cf6uqu8GSoiQt26qJpVR8J01Wee1YDHlfQ8P+OM2Rdzl9Bdm77d2Fdl5U
+35cpeniogHmzIYtwsegFTQIDAQABo0IwQDAdBgNVHQ4EFgQULEfZ+uKFk9G9c7+8
+Y0JnOTDhFQwwHwYDVR0jBBgwFoAUNQQu7tqwVgzhpoS5l8c8jVscMagwDQYJKoZI
+hvcNAQELBQADggEBABBvwCD959xZ3Gz82UPBNxVRoz5hfwg2rrb4JcZHbKfD4gfo
+M47YaRqHfecUW06WBmi4HgsLnimggjKQSiUDQGTuBZuotSpqJsj+7Y71MUYr8+CU
+UJ+ejs1Ic2S65+GIu0+KxdEpzMfMA/mYlQRKmiETJ5xZXqtsLETdZZVbHPZN0BnJ
+QKXeR6B8UCcw/XOtzc0TqQh3+V7gPx1NGe5y+E0xCwNgHtNGr2M4speAi2eb72qR
+vlxbVMpIvlrdZV17Gi4CF5bwGTSxoxiNIk4zTwMJakDoDcPM463HBVYShQBJqsau
+N3ySfzb5hhKWTBX665a0poEYTLlizf0GVfB5SLs=
 -----END CERTIFICATE-----
 """
 
 private let serverKey = """
------BEGIN RSA PRIVATE KEY-----
-MIIEpQIBAAKCAQEA0RAtEvag4c8c3sqjOY/wqzwJyd0KEmu6c1MkBYKIwD8cURVV
-/KWdCk/Jz+35bbIP534+wmSe7aEjU7mjndIGwQObW1odoIm326vk0qkSmHM4jtW7
-CSBygaqZjQfPsaYGZsbTpnqd7wRl+Dlr+uPGFbcbI6YhPez3FaJTcxbtVfQ8OVWa
-RnNVi1xKn34BdAwOm8srjwg27WZQLq7WdsTOPoq4XClTn5XVxEsd6v4j0dUDuoFy
-CwgMMR1QMbT+XQyK8DINs0cYcNlyx+OPBMywZFx34g40i0fJhvvb4kEQq/f+oUB7
-hch5HqIlA/g7FngKjF5QWCRVDCoE137UlgO68QIDAQABAoIBAEumodjh2/e6PYU1
-KHl0567e69/bF4Dw8Kg4pqlDwf5nF/UTVmk0+K25j5qpT3/tVin7mfQ3+vacP69V
-VqqOTJldl8Mnyd7E1v4rpoLAYZU+5HFzT9oOnsDjHetVr0dmf5yDSCVO64WJPuji
-xnskHxLOjoiI3jCNZh+y/KWB32IhdofSwBccw852JM2qC5l8vgE+sfjOeWXDiPRI
-YLlVRlxZFv7N2kn+EDnPEQ8m2OGYKvNzU0d9nz05NdkRXzMh9zegTTL4EQhTaMf0
-2AXy2ekKFVvWouV4y8QW1shz5Tun2y4ZQJnwiCyldED9sMxaziQxfdJ6N7f4+K5c
-Sh4+Ct0CgYEA7bGvY02jQfHcDPOjZ/xkXb98lr1uLGTSvwK3zw+rI0+nrUJwH6fB
-nSaXyWk059OqHTKPpa/d8DxFL2LP6vbvfTWCv9mnn61WWWDmP0Eo/k93XgWkmclb
-vQGfXV2wtCTnhz+iUSSJA8f8jZhCtOD6xa8pLsaYrGD6oR5wzfs0nysCgYEA4SoC
-/JWDMkw4mndI2vQ8GqDzBFtJCMr/dva7YGCGtbimDzxuOI68vW/y4X8Izg2i1hVz
-iKRYCI9KzRdQrQ7masZF2d4DeaqPeA72JN4hoUOw0TZjP8yD4ECifUt786ZNvV1n
-NlEhNb55zD73Cl6v0OJZkEfp1MC5ZQwLw7bMYFMCgYEAvNu5Z0WAuhzZotDSvQSl
-GnfTHlJU/6D8chhOw47Hg77+k4N+Yyh/hcXsRHP7PVfIinpp+FPMG91He2cfnKmn
-j+y8foMJ1K19NnbvesLjN20cgvAo4KhE4+AuJ5kRlZDdBXFiHubQltiHqlmYZu97
-USbjqe7Rz+UePnZZWtCF9xECgYEAjGODZTVbjdrUWAsT0+EAMKI1o3u/N8pKKkSA
-ZAELPPaaI1nMZ1sn9v179HkeZks+QjkxxfqiIQQm4WUuGhj2NZDWMJcql4tu1K6P
-bkFJuqDX+Dnu+/JqL0Jdjb2o1SvVwMIh/k3rZPUUP/LqWP7cpGLc8QbFlq9raMNv
-+mFZYJ0CgYEA5IQzp7SymcKgQqwcq5no2YOr76AykSCjOnLYYrotFqbxGJ19Cnol
-Z74Habxjv89Kc7bfIwbz/AolkhAS2y0CYwSJL4wZIUb8W2mroSmaHhsk3A4DMPBB
-wgSsdiBpixQqNDUAvnHc3FIyAGdpA73TJQrGY2F6QyQ9re3a/R8Dc3k=
------END RSA PRIVATE KEY-----
+-----BEGIN PRIVATE KEY-----
+MIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQDT/VX2vDPZ5EU3
+B2xpVlcfpabRW3yU+7ts/A73pngs2Ntw6M8YPHZ1GQxRe+djB3UH51Dd2r7m02Gk
+wugKMrOJ9zbQhAsmvW20m09YSflGi/47PZzCXPaxdH188xvh1XznS2rPHE7luJQ2
+G4RqdESLLB5neXODQPVUw41aMrJveFlNgHDi/qRwftRyF5pXadgTEfq0II7G4ZeR
+JJvkrFFdwzFs1YslKXEleMXgkUvgftwLl45VJGF0fIrO+ERFkuWwWeMfX4J/q6q7
+wZKiJC3bqomlVHwnTVZ57VgMeV9Dw/44zZF3OX0F2bvt3YV2XlTflyl6eKiAebMh
+i3Cx6AVNAgMBAAECggEABeEFNMoEYBsSBlQo4AT6dpC3/x9b5Z/Tk1KsMZcGxyue
+blOuhpwKpvWyX0Ih6R+oUbdLCVQLLkiXcSPMnpfIXd+iwAahboobVEJoYWjvxlKU
+kt0DD2rXlpVF86SkVc9/XsS7KhI9Z+aGCbgapH8j9X4cO79lAjNXemnP7jvwseGm
+HifmYsgzOqNoGIQsJS5diezmNHAmFeo0q577/fsQ0U1vnZkdTFUJF6ixPfQ5Wu/C
+O/E1pZnjLjFKXLwxTq59jW9ayKhGNMCnLME+ytScSAF5yNIr2bKjzlzaPP05//i2
+GXsfdn+3sjAsBAXB/XNWiNxDlIW+mzpb7KYZx3vs8QKBgQDx/EZmTnDiHw/nrjFw
+gRsC0bYQr7Xzw54gevooBFeJNfavvtlUcFNo3s9bFt4SGP7J6D4bQu0LtaOV/E6Y
+Z8CQCgbC6HXMpAWQqP4Q6vL2Hmi6YOh9ZWXyP3UrTjmwlolv0Z9BtLtx9/lAoz/W
+HwaMIIL8Mf2TH8YLaXV0zkCtqQKBgQDgRFYQrodrSzd3YBEr1GepntLYkwzvX+dV
+3qXk6mTeE58Fy3Yi7JEYcltB/Ynpin61bbpF9Fm72j9ZP1KwUjHMwFM8YiWFiq2l
+z/C2vGrhl2biK5FMCJ0FSxFWh0AtRoRsbjE0+wTz5a6XSzYV/lkf1f7qnpEJ1OkE
+VCWy/Go5BQKBgQDXlNrbNjLXILk4vDWLd1mrD03WW9QMEVCEu0q17/hUY2EcfTh9
+Q3zrxuSQ0DSThvUmx+LcwqkQb4gHjTowCO7C5EvVxOd9ovraP9PpGofNwJWoxcEa
+RgWm5eXq6Mv4iIR0vBnXnb4I8NR8Q3QRvJ5GpIhNSfQQ6JO1gwSDRWTV0QKBgCiZ
+NltevPUXcLsTkQg2M23papH6TOUon2BUJgQCyq/JLYiHBKPd34ZlbU/M7iJxB+8g
+JhBz46q2H7GwXO++cXV3e5n6yoAaUsUpl5H9GfwsxFs9N9hj1skWhdT8Nwn7Mg3P
+BOlqZ8MfWTPaUbg5izjQLI2clvUIrgYylYCJYVFhAoGBALgcxQRSj+Zi9Vao/Fgx
+3BfKMl88saz/77lbaHozTWi5W8FXPwZtDWfeAQexmYMZzzTh9bMQIoU2xL9RjlGU
+tcvJX6MnO8zV1RxhudwuWP2jXH3/B3jwfzh1D4srdOJO6zxBWVIHWCyZaR+4MQrk
+213NPjUNlKcqXrGNceXq+xu6
+-----END PRIVATE KEY-----
 """
 
 private let exampleServerCert = """
 -----BEGIN CERTIFICATE-----
-MIICnDCCAYQCAQEwDQYJKoZIhvcNAQELBQAwEjEQMA4GA1UEAwwHc29tZS1jYTAe
-Fw0yMjExMTQxMTE5MzRaFw0yMzExMTQxMTE5MzRaMBYxFDASBgNVBAMMC2V4YW1w
-bGUuY29tMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA5UPdc3MERjIj
-rKNMcsCJEPJtzFZG7T99rDaENxl5hF5TdlCuMfKQMyf4rmUk2KdQXduWDmP/9keO
-Btc3Hw9xm7mHn7UPbK0kHjlncqTCnjZzVQ2j0stg4Q0WjGeS0aB8k1AHPiBaOnvJ
-LYcBJrA8mZK+inEE0gWEJsODTM+bKb3+5I69qVoHAkU2tXTDV9g1YKfP4H3rufEg
-622AR1yAo8UxaGjY3amWps6XF/9R2iaSDAPLH1dCBw/YWrIH51n75S+n/H3Rz0+H
-/aT9Eze0M2F2Nj1cU8fVcbDNR0smssgXVmE2mvQ+OvbO0H7VTS1HK2q2aPOPkRWh
-yhFnOvPnbQIDAQABMA0GCSqGSIb3DQEBCwUAA4IBAQAWk5KHkWVsbXqz6/MnwxD5
-fn7JrBR77Vz3mxQO9NDKN/vczNMJf5cli6thrB5VPprl5LFXWwQ+LUQhP+XprigQ
-8owU1gMNqDzxVHn7By2lnAehLcWYxDoGc8xgTuf2aEjFAyW9xB67YP/kDx9uNFwY
-z+zWc6eMVr6dtKcsHcrIEoxLPBO9kuC/wlNY+73q04mmy9XQny15iQLy4sQT0wk4
-xV4p86rqDZcGepdV2/bLk2coF9cUOPOGwUBqEIc7n5GekC2WTXSnjOEK5c+2Wkbw
-Yt4jXnvsaQ8bwpchHfM1K3mLn+2rCEZ3F4E5Ug7DKebrwU4z/Nccf1DwM4pdI0EI
+MIIC5TCCAc2gAwIBAgIBATANBgkqhkiG9w0BAQsFADASMRAwDgYDVQQDDAdzb21l
+LWNhMB4XDTI0MDcyMDA2MjE0OVoXDTI1MDcyMDA2MjE0OVowFjEUMBIGA1UEAwwL
+ZXhhbXBsZS5jb20wggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQD3ETBd
+JjWyfnV48vsU2q+gH2Mt17iIOgOpvbQwOzgo5Q6li+zVXwgnPQAH23yHguY61NpS
+rEUO2+soCYE/19CRErTPJm1hWJpYXBdzR37vML6gkZ/Hi38JooHGzQgIvGatGnNt
+YUQKtDx290X0P1dlkbubqJYM0a3a6fzrft/aG/hRxVXAdIMqYS5U+G7jB7ruruD0
+/5l7l78y+hlV13OW5Mj9Hc9fA+mTrbGdWUtvcVBQ6EPkadmfp/bsK+A9JWEWo8eM
+rZk4MOWxxrjH1buvRuftDME04/zCj3oYM+Kq0QIDU6Pa/e1Kioc2Fcc2zrq7Gtvt
+awEvWxNwHbxAAL4fAgMBAAGjQjBAMB0GA1UdDgQWBBQw+Zc9GmEGX3mo96avqb64
+ebA4ATAfBgNVHSMEGDAWgBS7DtcKKCDw/Jy87h7R5aEnXeUMrDANBgkqhkiG9w0B
+AQsFAAOCAQEAA4mruZXxDLK+GevkI0bdptVGBx1C5O1b9W/qegAJJRZzRusK37mB
+WX6MhOYetoWqrXtJ+BVROM/OySWoUrcs9QlhVbrITff5d9ffsO8dSddgg7274UjG
+utlUgJa6UgkY9IfFZjOv05jbIt6wia1xUK22X3xdyFE+fsh3EJaKgeTlOO0FFK8d
+5kYiyWgU3oy6jicgijAndRHfV+NcIlWCcAlxjfkGoymUOF170v4upzJPSnDcKKXA
+5TW8D5sG0CAzFT5bEkIUZy7KQ7ojw6bhVrqsulC6QCCCQuYqDAqMnl74oOJAYdP3
+sDO1dOpe/tz7nf0CJNDGVFaY5d1xF9p8Cg==
 -----END CERTIFICATE-----
 """
 
 private let exampleServerKey = """
------BEGIN RSA PRIVATE KEY-----
-MIIEpAIBAAKCAQEA5UPdc3MERjIjrKNMcsCJEPJtzFZG7T99rDaENxl5hF5TdlCu
-MfKQMyf4rmUk2KdQXduWDmP/9keOBtc3Hw9xm7mHn7UPbK0kHjlncqTCnjZzVQ2j
-0stg4Q0WjGeS0aB8k1AHPiBaOnvJLYcBJrA8mZK+inEE0gWEJsODTM+bKb3+5I69
-qVoHAkU2tXTDV9g1YKfP4H3rufEg622AR1yAo8UxaGjY3amWps6XF/9R2iaSDAPL
-H1dCBw/YWrIH51n75S+n/H3Rz0+H/aT9Eze0M2F2Nj1cU8fVcbDNR0smssgXVmE2
-mvQ+OvbO0H7VTS1HK2q2aPOPkRWhyhFnOvPnbQIDAQABAoIBAQDk0+vAQ1hMx9ab
-hRHUpx8nbxDwFl0Mh4Zj0LX+WMrUt2EOglCbQcNzi73GMuWn6LdqNrV6/4yGv7ye
-T0iRE9UM3Qzk9s7CZb3a/OinoJMvXqGWjtqolp3HgkyzLt13pXsxfXr9I0Vrggm2
-Cz2248hYcAMGIu/wv9i66AGxNLVl3nzlo3K3J+6LwGYSrM5MMsN8p/RIc7RD30cg
-Qer6uiGdYemD2hbOuqcqImzhMLvoYn683uLOoDhiFLmAPIU+VxtHs2pMpp4ebjrl
-PpS8TtHnV85v/fhX6RE/jo5razdSU4LxW/p/fF5Zte+QR6FJgFFWaQvZd6/Vtuh6
-K0Hadt1xAoGBAP1fuBjUElQgWBtoXb422X2/LurnyHfNSqSDM3z8OiCSN08r5GmM
-ylWqh7k1sQWBzQ4OAsZcbwvrpvxMGYEd1K99LtUcM235WcKTomz7QRRUKG74tyFk
-VdCgcMF+q2DdBE+hlF08bTNk1dM6uPlinNiMklydhLFjjhLlfDkiteGTAoGBAOek
-LXqKMK4H5I1VQBgNKAv25tVItDabX5MPqhJVxmsvbNNTh/pfaNW7ietZNkec8FXs
-UtS2Hv2hwNMVSb8l9bk96b2x9wiww2exI4oWKjKkJrSVsIcWc4BgeZ2xUtnV6QR5
-XSNm7D4E11KhuHPbB01cAZeC0Cf/4rTZ5ERhULL/AoGAS/UpHJBfGkdEAptsFv0c
-gH0TFKr9xySNLvqCMgLvbhpHaH2xEQ97DOl9nMGC2zLJhWAf5tWJGNrBibtKnhGS
-VDXEF3FH3b018oYN2HwOS4jbQkFfrSwGKfAfPXK67+PySekXsEfQOOsOyy88is7M
-VIL30boLMJ621eVkM0C7o+8CgYBLiiK6n24YksJZxL9OGJxCqpXEYB1E4Y5davJP
-YGGAesrGb6scXxjU+n+TnFgzKl7F5ndsnqeklqdHLt4J09s6OZKMJgklcF+I5R9t
-3KSONzHYGiijJRMtfkiqwDUAjN2cc+eHr/zCjNmbPNnmDjtnYuWx/xrasHvB9nyW
-QBYNCQKBgQCDqdvchLcreSdbXKr6swvBz8XxzCaEParbm7iOvdLlv93svvCaHgI8
-6E+FlXk68Qc2Dj8de/xEnl/OonNQRgIQh7czBJmYP8+TCPECm8fvv2TsddNvjTmF
-TTx8wf9gixHffRtXZ4ILrP1sX7c4if3bfaMxKz+0ZyfzWZry8qZtVQ==
------END RSA PRIVATE KEY-----
+-----BEGIN PRIVATE KEY-----
+MIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQD3ETBdJjWyfnV4
+8vsU2q+gH2Mt17iIOgOpvbQwOzgo5Q6li+zVXwgnPQAH23yHguY61NpSrEUO2+so
+CYE/19CRErTPJm1hWJpYXBdzR37vML6gkZ/Hi38JooHGzQgIvGatGnNtYUQKtDx2
+90X0P1dlkbubqJYM0a3a6fzrft/aG/hRxVXAdIMqYS5U+G7jB7ruruD0/5l7l78y
++hlV13OW5Mj9Hc9fA+mTrbGdWUtvcVBQ6EPkadmfp/bsK+A9JWEWo8eMrZk4MOWx
+xrjH1buvRuftDME04/zCj3oYM+Kq0QIDU6Pa/e1Kioc2Fcc2zrq7GtvtawEvWxNw
+HbxAAL4fAgMBAAECggEAa4Ti976forCO3dZcNyyePoA6UKJu73XTz0Q7zLuQypc4
+Qwn8RLUJHs5Ym5LEhFKOb+7enhjTKs7A7RkJ0udQIDo0TmvqMPF2pdpZ6DSmhGBB
+rS4CizvlSqOad5IVm3ul8VIwBltMWZ4rLBibLFp+ZbjAnNKGjkiXsMugYjnCIC+u
+AZNXx9aWaULosOjyieTIm0Jx488kMZubQ/eHsyzM2vOFubvFkGbZIou0Tbz3Y4sn
+IZ2KysjCplQXIoW61I8GZQ9AUxJwGI+A9B6t8PZRYkRXh/JItxzGShWkqbfO6fm5
+KY/W5huYd8zSlfDQdX1sOdG7AvgMQocRvFGGkL+VGQKBgQD79OEldzA5SiLEN05m
+YwnbLYqe1NZaFAcWTvizQ4yM1UlzL4YT0lihRwvgExdkjfxG9SBWOkKOj9wS4wWI
+cynCKcCVqZwrNWC7Hjh9zF2UTIqMnvTIkgcQNnUoLD/EHRwynaJu3uGJ7zQSm8OW
+YHZYFqANBeHdy2HCNXCVWYcSCQKBgQD7CDjcSqnoeo5rA92tVQi+dSVX0qTkXP0Y
+biS45m+TudyREJk7wbIRZOQkvl175nBAMxB0jZUY4pPjy/9cigjg9ysNo5U9o6vQ
+OjPq9ug/XUNLb/iBIXJ1513szcbLOg7GfZBZxf6gpDBV3wGdNxuLR3dgXPtMJTcZ
+6mWtVyG45wKBgG1Qd4abya5xfdgH7tn5SkNv5N64lS+L1O9W4AxW2FoaB74T9mYP
+Myj1/C8B2XplJ/lZTOhcapgTznXbTYFABfQZzVahugH9LLTCjdv2mPtIoiwUozuK
+L8pW9kmxuRDLWHqVbe4QgWYpBDy2nAtYWsApQNOlo7NpFEcAjJIfv7yRAoGAdaZ9
+16jahs1hkAAo1cKjrXeDv+xGQYzfRaLhbRL6uhaCrC5GUr5N8V8CmcHqWFZAx6Xn
+EWkFtrsCKuIB1WIQFTdiIytGl7Pso1OT34xGfuP71RAuPH7rgRylZYVvpPGZ2Sci
+DyD/XFR3Hte8ju0z6IvfA/ewAxjyASBc9OeAM4MCgYEAmhGmc0EpFxcR8OVbFDuG
+3ZITjf7riyJDHai1/5hTaoJCH3OFEh3ztShY3Oos3P2IiApNbAmy8qrybxTa2KSJ
+NhAe6b+3E9TF5JnrfUPGNNhNKuqtV1eGeujhT7Le2JHk8nB37ehzxlm9eniGiL6C
+e+iul7j1acy9WAzAMM2yolU=
+-----END PRIVATE KEY-----
 """
 
 private let clientCert = """
 -----BEGIN CERTIFICATE-----
-MIICmjCCAYICAQEwDQYJKoZIhvcNAQELBQAwEjEQMA4GA1UEAwwHc29tZS1jYTAe
-Fw0yMjExMTQxMTE5MzRaFw0yMzExMTQxMTE5MzRaMBQxEjAQBgNVBAMMCWxvY2Fs
-aG9zdDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAOMSDHjt5P3MOmtw
-atYP0PnCfr2sxsMd1uxKeb+x5mpYuckcsnm1UR0LcBCizwBZ7yYAZQkFyK7vdJge
-Hv9P3Rki6+6Jj/ngLdpOirtUcOfnfzbdlA2k6qJtY6G+ZKyczDICWaZHzNRycDkL
-Yv4kzUT8PynIRIK/LPyXQa+tGty9+G2exVPdpzKpCgE8fKd8FeCOLW06Z0RsP0FS
-ySPSJxdDq0BRbfurhplhawh7uJ+7IoVfdWV2wwDvLztCEXHn2iiNpyzIixYapnVB
-PX1MXelsPRJaa9EKwOiqJB5ZcV9JWk9wa4W7mJrRfFTRh/9HRsoXIAaJPIqjTmqI
-ffat/1cCAwEAATANBgkqhkiG9w0BAQsFAAOCAQEAWpPkMzjKLAnxbfyasR6cPk3t
-ZanuBQ9RqVJw2gPem+SmSpvyPZ3CJTi0ZeUbPRSJ2W8YHyogaA5XShZSm8JJQKye
-TNYqUewVbU18OwVco0l7Wc8R1iiZgYEUcvMF2f/EWEMoCTmS3JlpbLl7LmdTy7iQ
-gIrR+iQ649nLw1T4Q5kp7zxjI6WJ3eZVNUjlTrUzKapSY4Tm2/+GafD+WNVRRACh
-Y9VNkaQ6qYy4SaLw6+bX2YdbDhIi275vAONHIZcAsMt6/aLJzKgfTxRqqmEvmJdQ
-KSVRRaSKZ/qe9UBdl4oFn1wupFAoNDQWkXT/Q3kVhxVXZ8ZE+ylZnuWcj3CHvQ==
+MIIC4zCCAcugAwIBAgIBATANBgkqhkiG9w0BAQsFADASMRAwDgYDVQQDDAdzb21l
+LWNhMB4XDTI0MDcyMDA2MjE0OVoXDTI1MDcyMDA2MjE0OVowFDESMBAGA1UEAwwJ
+bG9jYWxob3N0MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAyk9LkZPl
+QFsEqRpL7G3sodr994+S6c364WajthRUQsv69lJOurmEnA/JDekcLUEcVySOI5wR
+bmJ0hKgvfsmTRMx1v5NfOOAA/y9AiNZcpd/HLzaKGEYtqmw26ojcv5GYm9yqXMT2
+naw8V5l2GHVoQQdXqCRzd5CRWJl9ze1kqMVNYnex/EKKQDXv7kfGt2raB3WF7S4k
+EH2nahzjjoQ63/UY4IQaKiNxbalC5VwIuu9k2Qm+giWKyPJ+IxFpXM+2OcQoSZKe
+vMj1UeAhinxXz6/HMshA7RiKFfRmwVasEd12v4JzPLOa3LEp6y+e0ibyB0RzSLyG
+YzCqNI/hbIZrGQIDAQABo0IwQDAdBgNVHQ4EFgQUpPAVeogyMNdXKD/TazJFsgDx
+UKIwHwYDVR0jBBgwFoAUuw7XCigg8PycvO4e0eWhJ13lDKwwDQYJKoZIhvcNAQEL
+BQADggEBADqEVyhg7bA26yAhfv90754i1izMpjChd7yH3zgE6/BuaJ+dNoxY4LOx
+p5EWatGGQkhI+mPuaVQ39T9q+S6TW0icjAsXZxXnpK1usVmfmb8UJhvnEwNmP7n8
+SMcYPGEUGPdZjjb86/NXwgp9iOcwRdPzkltqL+n1stI+VoYNfhBkhkUF68+XH67K
+1UkjzxUKd0vJqiv0sSOzHuAyoiwgJ93FgTk90kp5jW2lru3mYdSoXKbMmxKkSmyU
+pgMctshh1Ps9Ts1yrng/w542MG1XrPyhaLBwGH3E2OuHZHV4M0UvJ2XHjnPthASN
+PFHeEB8Jed4SOWlxjQn5sfF3GzIznWM=
 -----END CERTIFICATE-----
 """
 
 private let clientSignedByOtherCACert = """
 -----BEGIN CERTIFICATE-----
-MIICoDCCAYgCAQEwDQYJKoZIhvcNAQELBQAwGDEWMBQGA1UEAwwNc29tZS1vdGhl
-ci1jYTAeFw0yMjExMTQxMTE5MzRaFw0yMzExMTQxMTE5MzRaMBQxEjAQBgNVBAMM
-CWxvY2FsaG9zdDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAOMSDHjt
-5P3MOmtwatYP0PnCfr2sxsMd1uxKeb+x5mpYuckcsnm1UR0LcBCizwBZ7yYAZQkF
-yK7vdJgeHv9P3Rki6+6Jj/ngLdpOirtUcOfnfzbdlA2k6qJtY6G+ZKyczDICWaZH
-zNRycDkLYv4kzUT8PynIRIK/LPyXQa+tGty9+G2exVPdpzKpCgE8fKd8FeCOLW06
-Z0RsP0FSySPSJxdDq0BRbfurhplhawh7uJ+7IoVfdWV2wwDvLztCEXHn2iiNpyzI
-ixYapnVBPX1MXelsPRJaa9EKwOiqJB5ZcV9JWk9wa4W7mJrRfFTRh/9HRsoXIAaJ
-PIqjTmqIffat/1cCAwEAATANBgkqhkiG9w0BAQsFAAOCAQEAIqGPpw/m3oIb+Ok7
-eZCEph/IcEwvkJXAFYeYjQHDsK1EW/HNoyjKKU3CaLJuWtvNbW+8GpmiVdAcO0XS
-RN+xwDOLmB+9Ob70tZRsR/4/695WkkCm/70Y89YqTq3ev86vZmPWBGZsdXB/rvfs
-sJbEkNeDRAquEbVQ3K8qmG7w8oC+VdzdQfQHY6hdkzsb0Q99aPASwGjxPVDz12Tb
-v9g9f9yVwI+vxxabHr4nvKJ/GfuHRzG2eSW2TNBY/Kxp10+lCdMfbPq2p0LsV4eZ
-eHPCFqiBe6CK80Pdpy7CNCPBBvGkGb7nfxBi4/tNVDgMlOQy6pA3PLjib8NLMCIA
-5iUEvw==
+MIIC6TCCAdGgAwIBAgIBATANBgkqhkiG9w0BAQsFADAYMRYwFAYDVQQDDA1zb21l
+LW90aGVyLWNhMB4XDTI0MDcyMDA2MjE0OVoXDTI1MDcyMDA2MjE0OVowFDESMBAG
+A1UEAwwJbG9jYWxob3N0MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA
+yk9LkZPlQFsEqRpL7G3sodr994+S6c364WajthRUQsv69lJOurmEnA/JDekcLUEc
+VySOI5wRbmJ0hKgvfsmTRMx1v5NfOOAA/y9AiNZcpd/HLzaKGEYtqmw26ojcv5GY
+m9yqXMT2naw8V5l2GHVoQQdXqCRzd5CRWJl9ze1kqMVNYnex/EKKQDXv7kfGt2ra
+B3WF7S4kEH2nahzjjoQ63/UY4IQaKiNxbalC5VwIuu9k2Qm+giWKyPJ+IxFpXM+2
+OcQoSZKevMj1UeAhinxXz6/HMshA7RiKFfRmwVasEd12v4JzPLOa3LEp6y+e0iby
+B0RzSLyGYzCqNI/hbIZrGQIDAQABo0IwQDAdBgNVHQ4EFgQUpPAVeogyMNdXKD/T
+azJFsgDxUKIwHwYDVR0jBBgwFoAUNQQu7tqwVgzhpoS5l8c8jVscMagwDQYJKoZI
+hvcNAQELBQADggEBAEZ0tQZ5tvRAbahuIgAuxdD1mk/cP4frxlrYpBxjfS5tZZwI
+lVYlrRbGeg/Ywqks51f5dvXZp6PxnOAPuqO86id2VQLYbTGRPa3cTFJcexmriBZc
+Cpeh+o1bkvG6RrHNVwdf8kBc45Nxqk3KJOTs6r4YdzagqEm914utKOHDb7LnTthO
+50ipeebNF77S3eFfGId3xlnvgnyKkybUmRudphFm2ptiQ4vxrLND+BOohtdlY2IL
+ABAwUVXzh4nygDo1jMNEOevSq5iuM7Hg0rGlkx1gZsFPAq3kkDnd0x02dxbnRWjK
+FxGhqC7H0fkzy84mf1malmWB9ghbJQGLHWEygS0=
 -----END CERTIFICATE-----
 """
 
 private let clientKey = """
------BEGIN RSA PRIVATE KEY-----
-MIIEpQIBAAKCAQEA4xIMeO3k/cw6a3Bq1g/Q+cJ+vazGwx3W7Ep5v7Hmali5yRyy
-ebVRHQtwEKLPAFnvJgBlCQXIru90mB4e/0/dGSLr7omP+eAt2k6Ku1Rw5+d/Nt2U
-DaTqom1job5krJzMMgJZpkfM1HJwOQti/iTNRPw/KchEgr8s/JdBr60a3L34bZ7F
-U92nMqkKATx8p3wV4I4tbTpnRGw/QVLJI9InF0OrQFFt+6uGmWFrCHu4n7sihV91
-ZXbDAO8vO0IRcefaKI2nLMiLFhqmdUE9fUxd6Ww9Elpr0QrA6KokHllxX0laT3Br
-hbuYmtF8VNGH/0dGyhcgBok8iqNOaoh99q3/VwIDAQABAoIBAQC5gA4mYJoY6FW1
-XcI5m/QphcWKaHJ8BY2FvZXWj5vftxoXfNUk7oYURzrGrGqVK+Nd1Sa1Bz+aAc7r
-UngaNQE3vrqlRUYUaRqsZEubm/Ec0pavmLaRqu9vwBOLmAGgrft2w0q/t5pS2CZr
-w6ycWC5FNBjZplypv0oeE+c6gB0YxKJ2mjKEYHWOop+uBPql2G6TfCeu4mMekZPH
-cHbMuMlBPN23HT7BmGCvk1YSaGbMt0iCTM0zThfe53AtLSVKn33szHm/XjLYJYGM
-7N+SttwM+O88diFShWHUHmWsy5Lv0Mrkw3NRz37yQ8Uh1fJ0TMeLZEIzKSLrI8lv
-XrBVE89ZAoGBAP8bBSytb6Aof6p1noIx8nu31d5/5mvRSUxoF1Lu/B/EeaNTGXxD
-HvJEi4Lh/txm6/3IeC5gfJ0jExxoWyTD3wLqVvmf+FEMntXqnATC832uw3vkxZpd
-E/MldDHE4UkWGBUvii7JN1fisyyZKLUu517crebJthpwk5Sf/1FgK1SbAoGBAOPd
-3VTQ7uaD2zPn4QM9KZsFJ+7cS7l1qtupplj9e12T7r53tQLoFjcery2Ja7ZrF3aq
-y07D2ww8y8v1ShxqTgSOdeqPCX1a4OS7Z93zsy58Jv3ZcXbfbSGiLbpoueQJbUZ0
-vKlDIf4uHn78fz8WIbe87UwKneKnaRrO64DtHQX1AoGBAIH11vYCySozV46UaxLy
-tRB3//lg+RcWQJwvLyqt2z2nzzv4OrSGUT6k0tnzne3UdQcN2MPvnaxD0RmYxE3/
-hx4qGfMDnvJTVput8JuwYXE21hnI2y4fmuk0vHQaU5bzLYOle2UIVyxrrlHbGNTs
-tywpimJXgnEHxvdhZyWis5BfAoGBAN45P2M6J+KzcRGb8DuiaHMAgkNWoJsMAEcd
-mldrTeajINCsGeHtycyTpi/4tw0+P7HBO2ljZLr4h6AvZcl0ewXCkYjhWlXgTTeE
-9PTmeDa7aaNjbl6J4vpMGeCTxcZ40xNFQcCo8fvbqm4ZfVdfFB8Gpz3jlLq4na5B
-YjdoB0gJAoGAZCK3JbIN56KnmyENuZ6szWTZwkCMZq3kPVKBK8LAITVLVxg7Emjs
-GyTU+JhMx9Hk2tU/tftM/dTZ2TRRMwmPbZNadtkQdDgsXDhfkrW9JmVewx4ECCcI
-gBfWFOoABVTmVM9oNc74FeWu3nDjqGix5ZJ8+Zjjr8wUEcrU2TPZKn4=
------END RSA PRIVATE KEY-----
+-----BEGIN PRIVATE KEY-----
+MIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQDKT0uRk+VAWwSp
+Gkvsbeyh2v33j5LpzfrhZqO2FFRCy/r2Uk66uYScD8kN6RwtQRxXJI4jnBFuYnSE
+qC9+yZNEzHW/k1844AD/L0CI1lyl38cvNooYRi2qbDbqiNy/kZib3KpcxPadrDxX
+mXYYdWhBB1eoJHN3kJFYmX3N7WSoxU1id7H8QopANe/uR8a3atoHdYXtLiQQfadq
+HOOOhDrf9RjghBoqI3FtqULlXAi672TZCb6CJYrI8n4jEWlcz7Y5xChJkp68yPVR
+4CGKfFfPr8cyyEDtGIoV9GbBVqwR3Xa/gnM8s5rcsSnrL57SJvIHRHNIvIZjMKo0
+j+FshmsZAgMBAAECggEAChIgMIUglX9qaBoUYza2H7kBaTZLEYs0xM1pkx3oM2KI
+0cP3YhVr1+Jzx1beNZT/Con57uUdG4X4Z6WhJqI1a+hzKJgXFoNVHV7eayljl4jA
+FzYzceGlco08PxjUx08BvimCDAfsgF0TbGdp4O4QRGNvjkHRQyeNyq+qWurJNtrz
+1SK7ziypzDgnE8Q3hJadO9yWozmZ14OqCPR4hrzGgFPz5zz17gGapi58Ux+R1lE3
+n9zpQw3g73xpO4y1KBZsElBkgHHAZBgScS8+Lm/N2NUwwxZRFWkDWPLJKqtfY8TP
+UaI+cKboj6e3LDlsfW3n+BoTnVt3rAnOThU6AY9FYQKBgQDpB0kfs36LgiNSwAcl
+sVElULUmvhQvTso8bmSsAM9+19/C8EhelWKXkHtEluv+C/8uMrEM0Oon02amO78G
+ojnaNPFkSKuD0/ieGB6NbOyiivqq+MroEMxmP5zZB6fTLD7gYnbRvcZOMwVGEYr3
+IH80Nldyg8qS5buCjAJKMkcx4QKBgQDeQMqBseidENfikJAJ05qoF4J377pEFMpv
+b9r7GutXS/5JVrdypTg7SsNK7x06pcDF2h/VHVfpT6l36u+nVn/+MZz1jQBsvJ5G
+FE0rG6FScZlzqPhXTYDIgrEeXBMRqMKCl+Sxo7PGEjit9JwdSJxwnEI7/YGMu4oP
+WeJNadZQOQKBgQC5AzjSe49wMi3baG0ERykXCGIbVXTvLo4wtrElQyt7GOgmfOHo
+fpCdp2/HQFA0DtBWnJanh0MxxY1NVEhRewGVts0lAvFyJBzTYHD5dk7WqqNklU18
+PMIcOEtOoN5lMk7llieiVRsaQTXhsNH1bY6xJKf/WpW+VbmvZ3v17r9J4QKBgCgN
+3zEeggCEf5O6X+sVcqLmkcHRt2z7njwVGY71nAJAlNRcx3Tm58pvv+BqVYczRhmi
+JOiAgE5w5+bPtV8gDLMf99ydC72NBK02jW2+bgVwqaoZgqAhPJtmSkthZqJhGqT0
+gOeuzABfWk2nbtcsNd1pm9o55rYZOlln5ovzDgepAoGALJWZQKSclb0hGJLL9dNP
+JK7OpXUMcska+Jm98nlRUwh1gZP/SIfZlVaYC7sGKb4oyoZRwDIyJ9WT42LvVnMb
+fT4CWkkvhC8HPIQVlPS2Qtjf3vwQozzAZ1G6dc8qkOe/ypb9ouVqJn+HEdufaFpQ
+4/B7TZKyKxbAEIBm8M28cWg=
+-----END PRIVATE KEY-----
 """
 
 private let serverExplicitCurveCert = """
 -----BEGIN CERTIFICATE-----
-MIICEDCCAbYCCQDCeNe2vM7d6DAKBggqhkjOPQQDAjAWMRQwEgYDVQQDDAtleGFt
-cGxlLmNvbTAeFw0yMjExMTQxMTE5MzRaFw0yMzExMTQxMTE5MzRaMBYxFDASBgNV
-BAMMC2V4YW1wbGUuY29tMIIBSzCCAQMGByqGSM49AgEwgfcCAQEwLAYHKoZIzj0B
-AQIhAP////8AAAABAAAAAAAAAAAAAAAA////////////////MFsEIP////8AAAAB
-AAAAAAAAAAAAAAAA///////////////8BCBaxjXYqjqT57PrvVV2mIa8ZR0GsMxT
-sPY7zjw+J9JgSwMVAMSdNgiG5wSTamZ44ROdJreBn36QBEEEaxfR8uEsQkf4vObl
-Y6RA8ncDfYEt6zOg9KE5RdiYwpZP40Li/hp/m47n60p8D54WK84zV2sxXs7LtkBo
-N79R9QIhAP////8AAAAA//////////+85vqtpxeehPO5ysL8YyVRAgEBA0IABDmd
-3Pzv6HbsUTmNd7RljKbkYP+36ljl6qVZKZ+8m3Exq4DvtIzLKho/4NluAhWCsRev
-2pWTfEiqiYS/U40TnfQwCgYIKoZIzj0EAwIDSAAwRQIhAI+BpDBjiqZD7r5vhPrG
-TT9Kq9s4ekIc1/a4AoTioT8CAiAluJHscXt+vBcqEI9sH0wudusCdPJyLbvNtMZd
-wdduCw==
+MIICdjCCAhugAwIBAgIUEhFu+6dIztYWjsk3W4hNZytHsrEwCgYIKoZIzj0EAwIw
+FjEUMBIGA1UEAwwLZXhhbXBsZS5jb20wHhcNMjQwNzIwMDYyMTQ5WhcNMjUwNzIw
+MDYyMTQ5WjAWMRQwEgYDVQQDDAtleGFtcGxlLmNvbTCCAUswggEDBgcqhkjOPQIB
+MIH3AgEBMCwGByqGSM49AQECIQD/////AAAAAQAAAAAAAAAAAAAAAP//////////
+/////zBbBCD/////AAAAAQAAAAAAAAAAAAAAAP///////////////AQgWsY12Ko6
+k+ez671VdpiGvGUdBrDMU7D2O848PifSYEsDFQDEnTYIhucEk2pmeOETnSa3gZ9+
+kARBBGsX0fLhLEJH+Lzm5WOkQPJ3A32BLeszoPShOUXYmMKWT+NC4v4af5uO5+tK
+fA+eFivOM1drMV7Oy7ZAaDe/UfUCIQD/////AAAAAP//////////vOb6racXnoTz
+ucrC/GMlUQIBAQNCAASHfeOIZlrc+BdonIMCZ0FYirqKGwHyrPONGXeU5fbi8Qpp
+Znr4tkgcHXs0mm3LPAFtvxLAdI+UjtOlZd2b/fi/o1MwUTAdBgNVHQ4EFgQUzpDC
+GzftjDa6oEyeHawytyeJHKMwHwYDVR0jBBgwFoAUzpDCGzftjDa6oEyeHawytyeJ
+HKMwDwYDVR0TAQH/BAUwAwEB/zAKBggqhkjOPQQDAgNJADBGAiEAmxa4eQTCbKGt
+kZFhZjbgNJhstHPCRG1HeU4P7PgiTwMCIQC4fPOdsO4wptaBwadQQo7QvuDYaWYI
+nEyZf64SLpplsA==
 -----END CERTIFICATE-----
 """
 
 private let serverExplicitCurveKey = """
 -----BEGIN EC PRIVATE KEY-----
-MIIBaAIBAQQgBLTFlKchn4c+dQphsqJ2hWVpLPeRQ0opnSwvRsH+63iggfowgfcC
+MIIBaAIBAQQgA2RFjdUNwBPcPA49GcuVkrij4TYwci+/M5t9FH7uT4CggfowgfcC
 AQEwLAYHKoZIzj0BAQIhAP////8AAAABAAAAAAAAAAAAAAAA////////////////
 MFsEIP////8AAAABAAAAAAAAAAAAAAAA///////////////8BCBaxjXYqjqT57Pr
 vVV2mIa8ZR0GsMxTsPY7zjw+J9JgSwMVAMSdNgiG5wSTamZ44ROdJreBn36QBEEE
 axfR8uEsQkf4vOblY6RA8ncDfYEt6zOg9KE5RdiYwpZP40Li/hp/m47n60p8D54W
 K84zV2sxXs7LtkBoN79R9QIhAP////8AAAAA//////////+85vqtpxeehPO5ysL8
-YyVRAgEBoUQDQgAEOZ3c/O/oduxROY13tGWMpuRg/7fqWOXqpVkpn7ybcTGrgO+0
-jMsqGj/g2W4CFYKxF6/alZN8SKqJhL9TjROd9A==
+YyVRAgEBoUQDQgAEh33jiGZa3PgXaJyDAmdBWIq6ihsB8qzzjRl3lOX24vEKaWZ6
++LZIHB17NJptyzwBbb8SwHSPlI7TpWXdm/34vw==
 -----END EC PRIVATE KEY-----
 """
 
